added dependencies

+++ src/airplane/g0/PlayerNew.java
@@ -87,7 +87,7 @@ public class PlayerNew extends Player {
 	   
 		 Plane p = planes.get(planeNumber);
 		 double initialBearing = p.getBearing();
-		 double minDistance = Double.MAX_VALUE;
+		 double minDistance = Integer.MAX_VALUE;
 		 double bestDirection = 0;
 		 double toReturn = initialBearing;
